requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.25
llama-index==0.11.18
llama-index-core==0.11.18
llama-index-readers-file==0.2.1
